+++ process_alignments_mouse.pl
@@ -588,6 +588,8 @@ if($ug){
     }
 }
 
+my $hasPair = 0;
+
 if($pair){
     `/bin/mkdir -m 775 -p $output/variants/snpsIndels`;
     `/bin/mkdir -m 775 -p $output/variants/snpsIndels/mutect`;
@@ -615,6 +617,8 @@ if($pair){
 	if($data[0] =~ /^NA$/i || $data[1] =~ /^NA$/i){
 	    next;
 	}
+        ## This means there really is a sample pair, haplotect should run.
+        $hasPair=1;
 
 	my $ran_mutect = 0;
 	my $mutectj = '';
@@ -779,7 +783,7 @@ if($pair){
         `/bin/mkdir -m 775 -p $output/variants/copyNumber/facets/$data[0]\_$data[1]\_facets/tmp`;
         my $facetsSETUP_jid = '';
         my $facets_setup = 0;
-        if(! -e "$output/progress/$pre\_$uID\_$data[0]\_$data[1]\_facets_SETUP.done" || $ssfj ) {
+        if($hasPair && (! -e "$output/progress/$pre\_$uID\_$data[0]\_$data[1]\_facets_SETUP.done" || $ssfj )) {
             my %stdParams = (scheduler => "$scheduler", job_name => "$pre\_$uID\_$data[0]\_facets_SETUP",  cpu => "4", mem => "5", job_hold => "$ssfj", cluster_out => "$output/progress/$pre\_$uID\_$data[0]\_facets_SETUP.log");
             my $standardParams = Schedule::queuing(%stdParams);
             my %addParams = (runtime => "30");
@@ -804,7 +808,7 @@ if($pair){
             $facetsSETUP_jid = "$pre\_$uID\_$data[0]\_$data[1]\_merge_counts_facets_SETUP";
         }
         ## now facets
-        if(! -e "$output/progress/$pre\_$uID\_$data[0]\_$data[1]\_facets_RUN.done" || $facets_setup){
+        if($hasPair && (! -e "$output/progress/$pre\_$uID\_$data[0]\_$data[1]\_facets_RUN.done" || $facets_setup)){
 
             my %stdParams = (scheduler => "$scheduler", job_name => "$pre\_$uID\_$data[0]\_$data[1]\_facets_RUN",  cpu => "3", mem => "2", job_hold => "$facetsSETUP_jid", cluster_out => "$output/progress/$pre\_$uID\_$data[0]\_$data[1]\_facets_RUN.log");
             my $standardParams = Schedule::queuing(%stdParams);
@@ -823,7 +827,7 @@ if($pair){
     close PAIR;
     my $facets_haplotect_jid = '';
 =begin FOR_FACETS
-    if(! -e "$output/progress/$pre\_$uID\_merge_facets_seg.done" || $facets_run){
+    if($hasPair && (! -e "$output/progress/$pre\_$uID\_merge_facets_seg.done" || $facets_run)){
         my $seg_outfile = "$output/variants/copyNumber/facets/$pre\_facets_merge_hisens.seg";
         if( -f "$seg_outfile"){
             unlink("$seg_outfile") or die "Cannot delete? $!";
@@ -842,7 +846,7 @@ if($pair){
 =end FOR_FACETS
 =cut
 
-    if(!-e "$output/progress/$pre\_$uID\_HAPLOTECT.done" || $ran_mutect_glob || $ran_hc){
+    if($hasPair && (!-e "$output/progress/$pre\_$uID\_HAPLOTECT.done" || $ran_mutect_glob || $ran_hc)){
         sleep(2);
         my $patientFile = "";
         if($patient){
@@ -861,7 +865,7 @@ if($pair){
 
 =begin FOR_FACETS
 
-    if(!-e "$output/progress/$pre\_$uID\_join_maf.done" || $haplotect_run || $facets_run){
+    if($hasPair && (!-e "$output/progress/$pre\_$uID\_join_maf.done" || $haplotect_run || $facets_run)){
         sleep(2);
 
         my %stdParams = (scheduler => "$scheduler", job_name => "$pre\_$uID\_join_maf", job_hold => "$facets_haplotect_jid", cpu => "4", mem => "8", cluster_out => "$output/progress/$pre\_$uID\_join_maf.log");
